requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
crispy-forms-foundation==0.7.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-cleanup==4.0.0
django-crispy-bulma==0.2
django-crispy-forms==1.9.0
django-datatable==0.3.1
django-datatables-view==1.19.1
django-heroku==0.3.1
django-session-timeout==0.1.0
django-tables2==2.3.1
django-tables2-bootstrap4==2.0.0b5
django-tables2-reports==0.1.3
django-use-email-as-username==1.1.1
filelock==3.0.12
gast==0.3.3
google-pasta==0.2.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
Markdown==3.2.1
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
path==13.2.0
Pillow==7.1.1
pip-review==1.1.0
protobuf==3.11.3
#psycopg2-binary 
public==2019.4.13
pyparsing==2.4.7
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
self==2019.4.13
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.25.9
virtualenv==20.0.18
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
zipp==3.1.0
